feat: migrate improved-webtools to CLI-first adapters
Why: Shared plugin policy now requires improved-webtools to treat the
standalone CLI as the product surface and reduce both the OpenCode
plugin and MCP server to thin adapters over one canonical execution
path.

Changes:
- Extracted the fetch/search logic from `src/index.ts` into shared
  operations plus a structured Bun bridge entrypoint, then reduced the
  plugin wrapper to OpenCode-only permission and metadata handling.
- Added a Typer-based CLI and FastMCP package under `mcp-server/src`
  that both delegate through the same bundled bridge artifact.
- Added a repo-local Wikipedia converter script, bundled the bridge for
  packaged CLI/MCP use, and updated env handling so cache paths,
  search config, and debug tool ids resolve per invocation.
- Reworked repo automation, package metadata, README flow, and tests so
  the CLI is documented first and `just --justfile justfile check`
  validates the new architecture.

Decisions:
- Chose a bundled Bun bridge inside the Python package instead of a
  second Python implementation so CLI, plugin, and MCP all share the
  same TypeScript core without depending on repo-root sources at
  runtime.
- Replaced the legacy `opencode run` integration test with adapter-level
  integration coverage because the old path depended on shared global
  OpenCode state and was failing on a readonly database outside the
  repo.
- Added repo-local `TMPDIR` and `UV_CACHE_DIR` exports so the canonical
  `just` workflow works in this workspace rather than relying on manual
  shell overrides.

Expected outcome: `improved-webtools` is available as a Typer CLI via
`uv run --directory mcp-server` or `uvx`, the plugin and MCP server are
thin adapters over the shared bridge, and `direnv exec . just --justfile
justfile check` passes.
