+++ .envrc
@@ -1,4 +1,7 @@
 source_up
+dotenv_if_exists .env
+export TMPDIR="/tmp"
+export UV_CACHE_DIR="/tmp/uv-cache-improved-webtools"
 
 # Required: SearxNG instance URL for web search
 # export SEARXNG_INSTANCE_URL="http://searxng.local"
