+++ mcp-server/src/improved_webtools/__init__.py
@@ -1,0 +1,6 @@
+"""CLI-first improved-webtools package."""
+
+from .cli import app
+from .server import mcp
+
+__all__ = ["app", "mcp"]
