+++ src/index.ts
@@ -1,861 +1,65 @@
 import { type Plugin, tool } from "@opencode-ai/plugin";
-import { getEncoding } from "js-tiktoken";
-import { createHash } from "node:crypto";
+
 import {
-  fetchArxivLibraryContent,
-  fetchGitHubContent,
-  fetchRedditPostMarkdown,
-  fetchYoutubeTranscriptMarkdown,
-  fetchWikipediaMarkdown,
-  isArxivLibraryUrl,
-  GITHUB_DOMAINS,
-  hostMatchesDomain,
-  REDDIT_DOMAINS,
-  type CommandExecutionResult,
-  type WebFetchDomainHandler,
-  type WebFetchHandlerResult,
-  WIKIPEDIA_DOMAINS,
-  YOUTUBE_DOMAINS,
-} from "./webfetch-handlers/index.ts";
-import { PASSPHRASE_WEB_SEARCH, PASSPHRASE_WEBFETCH } from "./passphrases.ts";
+  executeWebFetch,
+  executeWebSearch,
+  getToolDescriptions,
+  getToolIds,
+} from "./operations.ts";
 
-type SearxngResult = {
-  title: string;
-  url: string;
-  content: string;
-  engine: string;
-  category: string;
-  publishedDate: string | null;
-};
-
-type SearxngResponse = {
-  query: string;
-  number_of_results: number;
-  results: SearxngResult[];
-  answers: string[];
-  suggestions: string[];
-  unresponsive_engines: Array<[string, string]>;
-};
-
-type WebFetchCacheMode = "default" | "refresh";
-
-const SEARXNG_INSTANCE_URL = (process.env.SEARXNG_INSTANCE_URL ?? "").trim();
-const DEFAULT_TIMEOUT_MS = 15_000;
-const WEBFETCH_COMMAND_TIMEOUT_MS = 30_000;
-const WEBFETCH_WIKIPEDIA_CONVERT_TIMEOUT_MS = 120_000;
-const DEFAULT_LIMIT = 8;
-const MAX_LIMIT = 20;
-const MAX_OFFSET = 200;
-const MAX_PAGE_FETCHES = 20;
-const WEBFETCH_INLINE_TOKEN_LIMIT = 20_000;
-const WEBFETCH_CACHE_ENABLED = (process.env.WEBFETCH_CACHE_ENABLED ?? "1").trim() !== "0";
-const WEBFETCH_CACHE_DIR = (
-  process.env.WEBFETCH_CACHE_DIR ?? `${process.env.HOME ?? "/tmp"}/.cache/opencode-webfetch`
-).trim();
-const WEBFETCH_CACHE_TTL_DAYS = Number.parseInt(process.env.WEBFETCH_CACHE_TTL_DAYS ?? "90", 10);
-const WEBFETCH_CACHE_TTL_MS =
-  Number.isFinite(WEBFETCH_CACHE_TTL_DAYS) && WEBFETCH_CACHE_TTL_DAYS > 0
-    ? WEBFETCH_CACHE_TTL_DAYS * 24 * 60 * 60 * 1000
-    : 90 * 24 * 60 * 60 * 1000;
-const TOKEN_ENCODER = getEncoding("o200k_base");
-const ISSUE_REPORTING_HINT =
-  "If this looks like a technical tool-output issue, file it in ISSUES.md in this folder.";
-const REDDIT_APIFY_ACTOR = (process.env.REDDIT_APIFY_ACTOR ?? "spry_wholemeal/reddit-scraper").trim();
-const WIKIPEDIA_API_USER_AGENT = (
-  process.env.WIKIPEDIA_API_USER_AGENT ?? "opencode-improved-webfetch/1.0 (plugin)"
-).trim();
-const WIKIPEDIA_CONVERTER_SCRIPT = decodeURIComponent(
-  new URL("./scripts/wikipedia_html_to_markdown.py", import.meta.url).pathname,
-);
-const WEBFETCH_BASE_DESCRIPTION = "Use when you need to read a webpage URL as plain text content.";
-const WEBSEARCH_BASE_DESCRIPTION =
-  "Use when you need to search the web. Optional categories for narrowing only: news, it, npm, pypi, st, gh, hf, ollama, hn, science, arx, cr, gos, se, aa, lg. Use offset and num_results to paginate.";
-
-function envFlagEnabled(value?: string): boolean {
-  const normalized = (value ?? "").trim().toLowerCase();
-  return normalized === "1" || normalized === "true" || normalized === "yes" || normalized === "on";
+function buildLogger(client: {
+  app: {
+    log: (input: {
+      body: {
+        service: string;
+        level: "error" | "warn" | "debug" | "info";
+        message: string;
+        extra?: Record<string, unknown>;
+      };
+    }) => Promise<unknown>;
+  };
+}) {
+  return async (input: { message: string; extra: Record<string, unknown> }) => {
+    await client.app.log({
+      body: {
+        service: "improved-webtools",
+        level: "error",
+        message: input.message,
+        extra: input.extra,
+      },
+    });
+  };
 }
 
-const IMPROVED_WEBTOOLS_DEBUG_MODE = envFlagEnabled(process.env.IMPROVED_WEBTOOLS_DEBUG_MODE);
-const WEBFETCH_TOOL_ID = IMPROVED_WEBTOOLS_DEBUG_MODE ? "webfetch_debug" : "webfetch";
-const WEBSEARCH_TOOL_ID = IMPROVED_WEBTOOLS_DEBUG_MODE ? "websearch_debug" : "websearch";
-const WEBFETCH_DESCRIPTION = IMPROVED_WEBTOOLS_DEBUG_MODE
-  ? "Use only when explicitly debugging improved-webtools loading without shadowing the built-in webfetch tool. This debug-mode alias behaves the same as webfetch."
-  : WEBFETCH_BASE_DESCRIPTION;
-const WEBSEARCH_DESCRIPTION = IMPROVED_WEBTOOLS_DEBUG_MODE
-  ? "Use only when explicitly debugging improved-webtools loading without shadowing the built-in websearch tool. This debug-mode alias behaves the same as websearch. Optional categories for narrowing only: news, it, npm, pypi, st, gh, hf, ollama, hn, science, arx, cr, gos, se, aa, lg. Use offset and num_results to paginate."
-  : WEBSEARCH_BASE_DESCRIPTION;
-
-const NARROWING_CATEGORIES = [
-  "news",
-  "it",
-  "npm",
-  "pypi",
-  "st",
-  "gh",
-  "hf",
-  "ollama",
-  "hn",
-  "science",
-  "arx",
-  "cr",
-  "gos",
-  "se",
-  "aa",
-  "lg",
-] as const;
-
-const VALID_CATEGORIES = new Set<string>(NARROWING_CATEGORIES);
-function normalizeBaseUrl(baseUrl: string): string {
-  return baseUrl.replace(/\/+$/, "");
-}
-
-function parseLimit(limit?: number): number {
-  if (!Number.isFinite(limit)) return DEFAULT_LIMIT;
-  return Math.max(1, Math.min(MAX_LIMIT, Math.floor(limit ?? DEFAULT_LIMIT)));
-}
-
-function resolveOffset(offset?: number): { value: number; error?: string } {
-  if (offset === undefined) return { value: 0 };
-  if (!Number.isFinite(offset) || !Number.isInteger(offset) || offset < 0) {
-    return {
-      value: 0,
-      error: `Invalid offset: ${JSON.stringify(offset)}. Offset must be a non-negative integer.`,
-    };
-  }
-  if (offset > MAX_OFFSET) {
-    return {
-      value: 0,
-      error: `Invalid offset: ${JSON.stringify(offset)}. Maximum supported offset is ${MAX_OFFSET}.`,
-    };
-  }
-  return { value: offset };
-}
-
-function clampSnippet(text: string, maxLen = 280): string {
-  const clean = text.trim().replace(/\s+/g, " ");
-  if (clean.length <= maxLen) return clean;
-  return `${clean.slice(0, maxLen - 1)}…`;
-}
-
-function countTokens(text: string): number {
-  return TOKEN_ENCODER.encode(text).length;
-}
-
-function resolveWebFetchCacheMode(overwriteCache?: boolean): { value: WebFetchCacheMode } {
-  return { value: overwriteCache ? "refresh" : "default" };
-}
-
-type WebFetchCachePayload = {
-  url: string;
-  routeName: string;
-  sourceUrl: string;
-  content: string;
-  cachedAt: string;
-};
-
-type HttpMetadata = {
-  statusCode?: number;
-  contentType?: string;
-  contentLength?: number;
-};
-
-function webFetchCachePath(url: string): string {
-  const digest = createHash("sha256").update(url).digest("hex");
-  return `${WEBFETCH_CACHE_DIR}/${digest}.json`;
-}
-
-async function readWebFetchCache(
-  url: string,
-  overwriteCache: WebFetchCacheMode,
-): Promise<WebFetchHandlerResult | undefined> {
-  if (!WEBFETCH_CACHE_ENABLED) return undefined;
-  if (overwriteCache === "refresh") return undefined;
-  const path = webFetchCachePath(url);
-  const file = Bun.file(path);
-  if (!(await file.exists())) return undefined;
+function parseHttpUrl(value: string): URL | undefined {
   try {
-    const raw = await file.text();
-    const parsed = JSON.parse(raw) as Partial<WebFetchCachePayload>;
-    if (
-      parsed.url !== url ||
-      typeof parsed.routeName !== "string" ||
-      typeof parsed.sourceUrl !== "string" ||
-      typeof parsed.content !== "string" ||
-      typeof parsed.cachedAt !== "string"
-    ) {
+    const parsed = new URL(value.trim());
+    if (parsed.protocol !== "http:" && parsed.protocol !== "https:") {
       return undefined;
     }
-    const cachedAt = Date.parse(parsed.cachedAt);
-    if (!Number.isFinite(cachedAt) || Date.now() - cachedAt > WEBFETCH_CACHE_TTL_MS) {
-      await Bun.$`rm -f ${path}`.quiet();
-      return undefined;
-    }
-    return {
-      routeName: parsed.routeName,
-      sourceUrl: parsed.sourceUrl,
-      content: parsed.content,
-    };
+    return parsed;
   } catch {
     return undefined;
   }
 }
 
-async function writeWebFetchCache(
-  url: string,
-  result: WebFetchHandlerResult,
-): Promise<void> {
-  if (!WEBFETCH_CACHE_ENABLED) return;
-  if (result.routeName.includes("/binary")) return;
-  if (result.routeName.startsWith("arxiv/library")) return;
-  if (!result.content.trim()) return;
-  await Bun.$`mkdir -p ${WEBFETCH_CACHE_DIR}`.quiet();
-  const payload: WebFetchCachePayload = {
-    url,
-    routeName: result.routeName,
-    sourceUrl: result.sourceUrl,
-    content: result.content,
-    cachedAt: new Date().toISOString(),
-  };
-  await Bun.write(webFetchCachePath(url), JSON.stringify(payload));
-}
-
-function findWebFetchHandler(
-  handlers: readonly WebFetchDomainHandler[],
-  url: URL,
-): WebFetchDomainHandler | undefined {
-  return handlers.find((handler) =>
-    handler.domains.some((domain) => hostMatchesDomain(url.hostname, domain)),
-  );
-}
-
-async function runCommand(
-  args: string[],
-  timeoutMs = WEBFETCH_COMMAND_TIMEOUT_MS,
-): Promise<CommandExecutionResult> {
-  const proc = Bun.spawn(args, {
-    stdout: "pipe",
-    stderr: "pipe",
-  });
-
-  const timeout = setTimeout(() => {
-    proc.kill();
-  }, timeoutMs);
-  try {
-    const [stdoutText, stderrText, exitCode] = await Promise.all([
-      new Response(proc.stdout).text(),
-      new Response(proc.stderr).text(),
-      proc.exited,
-    ]);
-    return { stdoutText, stderrText, exitCode };
-  } finally {
-    clearTimeout(timeout);
-  }
-}
-
-async function fetchWebContentWithW3M(url: URL): Promise<CommandExecutionResult> {
-  return runCommand([
-    "sh",
-    "-lc",
-    `curl -sSL --compressed --max-time 30 ${JSON.stringify(url.toString())} | w3m -dump -T text/html`,
-  ]);
-}
-
-async function fetchHttpMetadata(url: URL): Promise<HttpMetadata> {
-  const result = await runCommand([
-    "sh",
-    "-lc",
-    `curl -sSIL --compressed --max-time 30 ${JSON.stringify(url.toString())}`,
-  ]);
-  if (result.exitCode !== 0) return {};
-
-  const headerBlocks = result.stdoutText
-    .split(/\r?\n\r?\n/g)
-    .map((block) => block.trim())
-    .filter(Boolean);
-  const finalHeaders = headerBlocks.at(-1) ?? result.stdoutText;
-  const lines = finalHeaders.split(/\r?\n/g);
-  const rawTrimmed = result.stdoutText.trim();
-
-  const statusLine = lines.find((line) => /^HTTP\/\d+(?:\.\d+)?\s+\d{3}\b/i.test(line));
-  const statusCode = statusLine
-    ? Number.parseInt(statusLine.replace(/^HTTP\/\d+(?:\.\d+)?\s+/, "").slice(0, 3), 10)
-    : /^\d{3}$/.test(rawTrimmed)
-      ? Number.parseInt(rawTrimmed, 10)
-      : undefined;
-  const contentTypeHeader = lines.find((line) => /^content-type:/i.test(line));
-  const contentLengthHeader = lines.find((line) => /^content-length:/i.test(line));
-  const contentType = contentTypeHeader?.replace(/^content-type:\s*/i, "").trim();
-  const contentLength = Number.parseInt(
-    contentLengthHeader?.replace(/^content-length:\s*/i, "").trim() ?? "",
-    10,
-  );
-
-  return {
-    statusCode:
-      Number.isFinite(statusCode) && (statusCode ?? 0) >= 100 && (statusCode ?? 0) <= 599
-        ? statusCode
-        : undefined,
-    contentType: contentType && contentType.length > 0 ? contentType : undefined,
-    contentLength: Number.isFinite(contentLength) && contentLength >= 0 ? contentLength : undefined,
-  };
-}
-
-function isPdfContentType(contentType?: string): boolean {
-  return contentType?.toLowerCase().includes("application/pdf") ?? false;
-}
-
-async function downloadPdfToTemp(url: URL): Promise<CommandExecutionResult> {
-  return runCommand([
-    "sh",
-    "-lc",
-    [
-      "set -eu",
-      "tmpdir=$(mktemp -d /tmp/webfetch-pdf-XXXXXX)",
-      'outfile="$tmpdir/document.pdf"',
-      `curl -sSL --compressed --max-time 30 -o "$outfile" ${JSON.stringify(url.toString())}`,
-      'printf "%s\\n" "$outfile"',
-    ].join("; "),
-  ]);
-}
-
-function formatArxivServiceMessage(input: {
-  url: URL;
-  statusCode?: number;
-  content: string;
-}): string | undefined {
-  if (!hostMatchesDomain(input.url.hostname, "arxiv.org")) return undefined;
-  const body = input.content.trim().toLowerCase();
-
-  if (input.statusCode === 429 || body === "rate exceeded." || body === "rate exceeded") {
-    return [
-      "arXiv API case: `429 Rate exceeded`.",
-      "Interpretation: this indicates arXiv server capacity pressure, not abusive request rate from your script.",
-      "Action: retry later with backoff; keep polite spacing (~3 seconds) between repeated API calls.",
-      "Reference: https://groups.google.com/a/arxiv.org/g/api/c/pNB3lnxf4mQ",
-    ].join("\n");
-  }
-
-  if (input.statusCode === 503) {
-    return [
-      "arXiv API case: `503 Service Unavailable`.",
-      "Interpretation: this is the excessive-use signal from arXiv API ops.",
-      "Action: reduce request frequency, use smaller slices/paging, and keep polite spacing (~3 seconds) between repeated calls.",
-      "Reference: https://groups.google.com/a/arxiv.org/g/api/c/pNB3lnxf4mQ",
-    ].join("\n");
-  }
-
-  return undefined;
-}
-
-function buildArxivFallbackUrl(url: URL): URL | undefined {
-  if (!hostMatchesDomain(url.hostname, "arxiv.org")) return undefined;
-  if (url.pathname !== "/api/query") return undefined;
-
-  const rawIdList = (url.searchParams.get("id_list") ?? "").trim();
-  if (rawIdList) {
-    const firstId = rawIdList
-      .split(",")
-      .map((item) => item.trim())
-      .find((item) => item.length > 0);
-    if (firstId) {
-      const normalizedId = firstId.replace(/^arxiv:/i, "");
-      const fallback = new URL("https://arxiv.org/");
-      fallback.pathname = `/abs/${normalizedId}`;
-      return fallback;
-    }
-  }
-
-  const rawSearch = (url.searchParams.get("search_query") ?? "").trim();
-  if (rawSearch) {
-    const fallback = new URL("https://arxiv.org/search/");
-    fallback.searchParams.set("query", rawSearch);
-    fallback.searchParams.set("searchtype", "all");
-    fallback.searchParams.set("source", "header");
-    return fallback;
-  }
-
-  return undefined;
-}
-
-function isLikelyArxivErrorBody(content: string): boolean {
-  const normalized = content.trim().toLowerCase();
-  return (
-    normalized === "rate exceeded." ||
-    normalized === "rate exceeded" ||
-    normalized === "service unavailable." ||
-    normalized === "service unavailable" ||
-    normalized === "service temporarily unavailable." ||
-    normalized === "service temporarily unavailable" ||
-    normalized.includes("service temporarily unavailable") ||
-    normalized.includes("too many requests")
-  );
-}
-
-async function formatWebFetchOutput(input: {
-  routeName: string;
-  sourceUrl: string;
-  content: string;
-}): Promise<string> {
-  const text = input.content.trim();
-  const prefix = [
-    `Tool passphrase: ${PASSPHRASE_WEBFETCH}`,
-    `Route: ${input.routeName}`,
-    `Source URL: ${input.sourceUrl}`,
-    ISSUE_REPORTING_HINT,
-  ];
-  if (!text) {
-    return [...prefix, "", "No readable text content extracted from this page."].join("\n");
-  }
-
-  const inlineReport = [...prefix, "", text].join("\n");
-  const tokenCount = countTokens(inlineReport);
-  if (tokenCount > WEBFETCH_INLINE_TOKEN_LIMIT) {
-    const outputPath = `/tmp/webfetch-${Date.now()}-${crypto.randomUUID()}.txt`;
-    await Bun.write(outputPath, inlineReport);
-    return [
-      ...prefix,
-      `Token count: ${tokenCount}`,
-      `Saved full content: ${outputPath}`,
-      `Full report exceeds inline limit (${WEBFETCH_INLINE_TOKEN_LIMIT} tokens).`,
-      "Use your normal file-reading tools to inspect the saved file.",
-    ].join("\n");
-  }
-
-  return [...prefix, `Token count: ${tokenCount}`, "", text].join("\n");
-}
-
-function augmentQueryWithDomains(query: string, domains: string[]): string {
-  if (domains.length === 0) return query;
-  // `site:` is query syntax forwarded to engines; behavior depends on engine support.
-  const siteClauses = domains.map((domain) => `site:${domain}`);
-  return `${query} (${siteClauses.join(" OR ")})`;
-}
-
-function mapRecencyToTimeRange(recency?: number): string | undefined {
-  if (!Number.isFinite(recency)) return undefined;
-  const days = Math.max(0, Math.floor(recency ?? 0));
-  if (days <= 1) return "day";
-  if (days <= 31) return "month";
-  return "year";
-}
-
-function buildQueryUrl(baseUrl: string, args: { query: string; timeRange?: string; pageNumber?: number }): URL {
-  const url = new URL(`${normalizeBaseUrl(baseUrl)}/search`);
-  url.searchParams.set("q", args.query);
-  url.searchParams.set("format", "json");
-
-  if (args.timeRange) {
-    url.searchParams.set("time_range", args.timeRange);
-  }
-  if (args.pageNumber && args.pageNumber > 1) {
-    url.searchParams.set("pageno", String(args.pageNumber));
-  }
-
-  return url;
-}
-
-function asResponse(input: unknown): SearxngResponse {
-  if (!input || typeof input !== "object") {
-    throw new Error("Invalid search API response: expected object.");
-  }
-  const response = input as Record<string, unknown>;
-
-  if (typeof response.query !== "string") {
-    throw new Error("Invalid search API response: missing query.");
-  }
-  if (typeof response.number_of_results !== "number" || !Number.isFinite(response.number_of_results)) {
-    throw new Error("Invalid search API response: missing number_of_results.");
-  }
-  if (!Array.isArray(response.results)) {
-    throw new Error("Invalid search API response: missing results array.");
-  }
-  if (!Array.isArray(response.answers)) {
-    throw new Error("Invalid search API response: missing answers array.");
-  }
-  if (!Array.isArray(response.suggestions)) {
-    throw new Error("Invalid search API response: missing suggestions array.");
-  }
-  if (!Array.isArray(response.unresponsive_engines)) {
-    throw new Error("Invalid search API response: missing unresponsive_engines array.");
-  }
-
-  for (const item of response.results) {
-    if (!item || typeof item !== "object") {
-      throw new Error("Invalid search API response: result item is not an object.");
-    }
-    const result = item as Record<string, unknown>;
-    if (typeof result.title !== "string" || typeof result.url !== "string") {
-      throw new Error("Invalid search API response: result item missing title/url.");
-    }
-    if (typeof result.content !== "string") {
-      throw new Error("Invalid search API response: result item missing content.");
-    }
-    if (typeof result.engine !== "string" || typeof result.category !== "string") {
-      throw new Error("Invalid search API response: result item missing engine/category.");
-    }
-    if (!(typeof result.publishedDate === "string" || result.publishedDate === null)) {
-      throw new Error("Invalid search API response: invalid publishedDate.");
-    }
-  }
-
-  return response as unknown as SearxngResponse;
-}
-
-function resolveCategory(category?: string): { bang?: string; error?: string } {
-  if (!category) return {};
-  const raw = category.trim().toLowerCase();
-  if (!raw) return {};
-
-  if (raw.startsWith("!")) {
-    return {
-      error:
-        `Invalid category: ${JSON.stringify(category)}. ` +
-        "Use bare category names only.",
-    };
-  }
-
-  const normalized = raw;
-  if (!VALID_CATEGORIES.has(normalized)) {
-    return {
-      error:
-        `Invalid category: ${JSON.stringify(category)}. ` +
-        `Allowed categories: ${NARROWING_CATEGORIES.join(", ")}.`,
-    };
-  }
-
-  return { bang: `!${normalized}` };
-}
-
-function formatResults(input: {
-  query: string;
-  category?: string;
-  response: SearxngResponse;
-  offset: number;
-  limit: number;
-}): string {
-  const top = input.response.results.slice(0, input.limit);
-  const start = top.length > 0 ? input.offset + 1 : 0;
-  const end = input.offset + top.length;
-  const total = input.response.number_of_results;
-
-  const lines: string[] = [];
-  lines.push(`Query: ${input.query}`);
-  if (input.category) {
-    lines.push(`Category: ${input.category}`);
-  }
-  lines.push(`Offset: ${input.offset}`);
-  lines.push(`Total results: ${total}`);
-  if (top.length > 0) {
-    lines.push(`Showing results: ${start}-${end} of ${total}`);
-  } else {
-    lines.push("Showing results: none");
-  }
-  lines.push(`Returned results: ${top.length}`);
-
-  if (top.length === 0) {
-    lines.push("No results returned.");
-    if (input.response.answers.length > 0) {
-      lines.push(`Answers: ${input.response.answers.map((x) => x.trim()).filter(Boolean).join(" | ")}`);
-    }
-    if (input.response.suggestions.length > 0) {
-      lines.push(`Suggestions: ${input.response.suggestions.join(", ")}`);
-    }
-    lines.push("");
-    lines.push("Continuation hints:");
-    lines.push("- Snippets are truncated. Use webfetch on a selected URL to read full page contents.");
-    lines.push(`- To continue pagination, rerun with offset ${input.offset + input.limit}.`);
-    return lines.join("\n");
-  }
-
-  lines.push("");
-  lines.push("Top results:");
-
-  for (const [index, result] of top.entries()) {
-    const rank = index + 1;
-    const title = result.title.trim();
-    const url = result.url.trim();
-    lines.push(`${rank}. ${title}`);
-    lines.push(`   URL: ${url}`);
-
-    if (result.publishedDate) {
-      lines.push(`   Published: ${result.publishedDate}`);
-    }
-
-    if (result.content.trim()) {
-      lines.push(`   Snippet: ${clampSnippet(result.content)}`);
-    }
-  }
-
-  lines.push("");
-  lines.push("Continuation hints:");
-  lines.push("- Snippets are truncated. Use webfetch on a selected URL to read full page contents.");
-  lines.push(`- To continue pagination, rerun with offset ${input.offset + top.length}.`);
-
-  return lines.join("\n");
-}
-
 export const ImprovedWebSearchPlugin: Plugin = async ({ client }) => {
-  const webFetchDomainHandlers: readonly WebFetchDomainHandler[] = [
-    {
-      name: "wikipedia",
-      domains: WIKIPEDIA_DOMAINS,
-      handle: async ({ url }) =>
-        fetchWikipediaMarkdown({
-          url,
-          runCommand,
-          converterScriptPath: WIKIPEDIA_CONVERTER_SCRIPT,
-          userAgent: WIKIPEDIA_API_USER_AGENT,
-          convertTimeoutMs: WEBFETCH_WIKIPEDIA_CONVERT_TIMEOUT_MS,
-        }),
-    },
-    {
-      name: "youtube",
-      domains: YOUTUBE_DOMAINS,
-      handle: async ({ url }) =>
-        fetchYoutubeTranscriptMarkdown({
-          url,
-          runCommand,
-        }),
-    },
-    {
-      name: "reddit",
-      domains: REDDIT_DOMAINS,
-      handle: async ({ url }) =>
-        fetchRedditPostMarkdown({
-          url,
-          runCommand,
-          fetchFallbackWithW3M: fetchWebContentWithW3M,
-          apifyActor: REDDIT_APIFY_ACTOR,
-        }),
-    },
-    {
-      name: "github",
-      domains: GITHUB_DOMAINS,
-      handle: async ({ url }) =>
-        fetchGitHubContent({
-          url,
-          runCommand,
-        }),
-    },
-  ];
-
-  const websearchTool = tool({
-    description: WEBSEARCH_DESCRIPTION,
-    args: {
-      query: tool.schema.string(),
-      category: tool.schema.string().optional(),
-      num_results: tool.schema.number().optional(),
-      offset: tool.schema.number().optional(),
-      recency: tool.schema.number().optional(),
-      domains: tool.schema.array(tool.schema.string()).optional(),
-    },
-    async execute(args, context) {
-      const baseUrl = SEARXNG_INSTANCE_URL;
-      if (!baseUrl) {
-        return [
-          `Tool passphrase: ${PASSPHRASE_WEB_SEARCH}`,
-          ISSUE_REPORTING_HINT,
-          "Web search is unavailable in this environment. Ask the user to check plugin configuration.",
-        ].join("\n");
-      }
-
-      const query = args.query.trim();
-      if (!query) {
-        return [
-          `Tool passphrase: ${PASSPHRASE_WEB_SEARCH}`,
-          ISSUE_REPORTING_HINT,
-          "Invalid search: query is empty.",
-        ].join("\n");
-      }
-
-      await context.ask({
-        permission: WEBSEARCH_TOOL_ID,
-        patterns: [query],
-        always: ["*"],
-        metadata: {
-          query,
-        },
-      });
-
-      const limit = parseLimit(args.num_results);
-
-      context.metadata({
-        title: `Web search: ${query.slice(0, 72)}`,
-        metadata: {
-          num_results: limit,
-        },
-      });
-
-      const timeoutSignal = AbortSignal.timeout(DEFAULT_TIMEOUT_MS);
-      const signal = AbortSignal.any([context.abort, timeoutSignal]);
-
-      const domains = (args.domains ?? [])
-        .map((domain) => domain.trim())
-        .filter(Boolean);
-
-      const category = resolveCategory(args.category);
-      if (category.error) {
-        return [
-          `Tool passphrase: ${PASSPHRASE_WEB_SEARCH}`,
-          ISSUE_REPORTING_HINT,
-          category.error,
-        ].join("\n");
-      }
-
-      const offset = resolveOffset(args.offset);
-      if (offset.error) {
-        return [
-          `Tool passphrase: ${PASSPHRASE_WEB_SEARCH}`,
-          ISSUE_REPORTING_HINT,
-          offset.error,
-        ].join("\n");
-      }
-
-      let effectiveQuery = augmentQueryWithDomains(query, domains);
-      if (category.bang) {
-        effectiveQuery = `${category.bang} ${effectiveQuery}`;
-      }
-
-      try {
-        const timeRange = mapRecencyToTimeRange(args.recency);
-        const neededCount = offset.value + limit;
-        const collected: SearxngResult[] = [];
-        let firstResponse: SearxngResponse | undefined;
-
-        for (
-          let page = 1;
-          page <= MAX_PAGE_FETCHES && collected.length < neededCount;
-          page += 1
-        ) {
-          const url = buildQueryUrl(baseUrl, {
-            query: effectiveQuery,
-            timeRange,
-            pageNumber: page,
-          });
-          const response = await fetch(url, {
-            method: "GET",
-            headers: {
-              accept: "application/json",
-            },
-            signal,
-          });
-
-          if (!response.ok) {
-            const body = await response.text();
-            await client.app.log({
-              body: {
-                service: "web-search-plugin",
-                level: "error",
-                message: `web_search request failed: HTTP ${response.status} ${response.statusText}`,
-                extra: {
-                  query,
-                  category: args.category ?? null,
-                  offset: offset.value,
-                  num_results: limit,
-                  status: response.status,
-                  statusText: response.statusText,
-                  responseBody: clampSnippet(body, 2000),
-                },
-              },
-            });
-            throw new Error(`search request failed (HTTP ${response.status})`);
-          }
-
-          const pageData = asResponse(await response.json());
-          if (!firstResponse) {
-            firstResponse = pageData;
-          }
-
-          const pageResults = pageData.results;
-          if (pageResults.length === 0) {
-            break;
-          }
-          collected.push(...pageResults);
-        }
-
-        if (!firstResponse) {
-          return [
-            `Tool passphrase: ${PASSPHRASE_WEB_SEARCH}`,
-            ISSUE_REPORTING_HINT,
-            "No results found.",
-          ].join("\n");
-        }
-
-        const windowedResults = collected.slice(offset.value, offset.value + limit);
-        const data: SearxngResponse = {
-          ...firstResponse,
-          results: windowedResults,
-        };
-
-        return [
-          `Tool passphrase: ${PASSPHRASE_WEB_SEARCH}`,
-          ISSUE_REPORTING_HINT,
-          "",
-          formatResults({
-            query,
-            category: args.category,
-            response: data,
-            offset: offset.value,
-            limit,
-          }),
-        ].join("\n");
-      } catch (error) {
-        const message = error instanceof Error ? error.message : String(error);
-        await client.app.log({
-          body: {
-            service: "web-search-plugin",
-            level: "error",
-            message: "web_search execution error",
-            extra: {
-              query,
-              category: args.category ?? null,
-              offset: offset.value,
-              num_results: limit,
-              error: message,
-            },
-          },
-        });
-        return [
-          `Tool passphrase: ${PASSPHRASE_WEB_SEARCH}`,
-          ISSUE_REPORTING_HINT,
-          `Search failed: ${message}`,
-        ].join("\n");
-      }
-    },
-  });
+  const toolIds = getToolIds();
+  const toolDescriptions = getToolDescriptions();
 
   return {
     tool: {
-      [WEBFETCH_TOOL_ID]: tool({
-        description: WEBFETCH_DESCRIPTION,
-        args: {
-          url: tool.schema.string(),
-          overwrite_cache: tool.schema.boolean().optional(),
-        },
-        async execute(args, context) {
-          const rawUrl = args.url.trim();
-          let parsed: URL;
-          try {
-            parsed = new URL(rawUrl);
-          } catch {
-            return [
-              `Tool passphrase: ${PASSPHRASE_WEBFETCH}`,
-              ISSUE_REPORTING_HINT,
-              `Invalid URL: ${JSON.stringify(args.url)}.`,
-            ].join("\n");
-          }
-
-          if (parsed.protocol !== "http:" && parsed.protocol !== "https:") {
-            return [
-              `Tool passphrase: ${PASSPHRASE_WEBFETCH}`,
-              ISSUE_REPORTING_HINT,
-              `Unsupported URL protocol for webfetch: ${parsed.protocol}`,
-            ].join("\n");
-          }
-
+      [toolIds.webfetch]: tool({
+        description: toolDescriptions.webfetch,
+      args: {
+        url: tool.schema.string(),
+        overwrite_cache: tool.schema.boolean().optional(),
+      },
+      async execute(args, context) {
+        const parsed = parseHttpUrl(args.url);
+        if (parsed) {
           await context.ask({
-            permission: WEBFETCH_TOOL_ID,
+            permission: toolIds.webfetch,
             patterns: [parsed.toString()],
             always: ["*"],
             metadata: {
@@ -866,134 +70,51 @@ export const ImprovedWebSearchPlugin: Plugin = async ({ client }) => {
           context.metadata({
             title: `Web fetch: ${parsed.hostname}${parsed.pathname}`.slice(0, 120),
           });
+        }
 
-          try {
-            const resolvedCacheMode = resolveWebFetchCacheMode(args.overwrite_cache);
-            const cacheKey = parsed.toString();
-            const useArxivLibrary = isArxivLibraryUrl(parsed);
-            const cached = useArxivLibrary
-              ? undefined
-              : await readWebFetchCache(cacheKey, resolvedCacheMode.value);
-            if (cached) {
-              return formatWebFetchOutput({
-                routeName: `${cached.routeName}/cache`,
-                sourceUrl: cached.sourceUrl,
-                content: cached.content,
-              });
-            }
-            const handler = findWebFetchHandler(webFetchDomainHandlers, parsed);
-            const fetched = useArxivLibrary
-              ? await fetchArxivLibraryContent({
-                  url: parsed,
-                  overwriteCache: resolvedCacheMode.value === "refresh",
-                })
-              : handler
-                ? await handler.handle({ url: parsed })
-              : await (async () => {
-                  const httpMetadata = await fetchHttpMetadata(parsed);
-                  if (isPdfContentType(httpMetadata.contentType)) {
-                    const downloadResult = await downloadPdfToTemp(parsed);
-                    if (downloadResult.exitCode !== 0) {
-                      throw new Error(
-                        `pdf webfetch failed (exit ${downloadResult.exitCode}): ${downloadResult.stderrText.trim()}`,
-                      );
-                    }
-                    const savedPath = downloadResult.stdoutText.trim();
-                    const details = [
-                      `Binary content detected: ${httpMetadata.contentType}.`,
-                      `Saved PDF: ${savedPath}`,
-                    ];
-                    if (httpMetadata.contentLength !== undefined) {
-                      details.push(`Content-Length: ${httpMetadata.contentLength} bytes`);
-                    }
-                    details.push(
-                      "PDF responses are downloaded directly to a temporary directory instead of being piped through w3m.",
-                    );
-                    details.push("Use your normal file-reading tools on the saved file.");
-                    return {
-                      routeName: "default/binary-pdf",
-                      sourceUrl: parsed.toString(),
-                      content: details.join("\n"),
-                    };
-                  }
-                  const defaultResult = await fetchWebContentWithW3M(parsed);
-                  if (defaultResult.exitCode !== 0) {
-                    throw new Error(
-                      `default webfetch failed (exit ${defaultResult.exitCode}): ${defaultResult.stderrText.trim()}`,
-                    );
-                  }
-                  const serviceMessage = formatArxivServiceMessage({
-                    url: parsed,
-                    statusCode: httpMetadata.statusCode,
-                    content: defaultResult.stdoutText,
-                  });
-                  if (serviceMessage) {
-                    const fallbackUrl = buildArxivFallbackUrl(parsed);
-                    if (fallbackUrl) {
-                      const fallbackResult = await fetchWebContentWithW3M(fallbackUrl);
-                      if (
-                        fallbackResult.exitCode === 0 &&
-                        fallbackResult.stdoutText.trim().length > 0 &&
-                        !isLikelyArxivErrorBody(fallbackResult.stdoutText)
-                      ) {
-                        return {
-                          routeName: "default",
-                          sourceUrl: fallbackUrl.toString(),
-                          content: [
-                            serviceMessage,
-                            `Fallback: loaded arXiv web page via w3m from ${fallbackUrl.toString()}.`,
-                            "",
-                            fallbackResult.stdoutText,
-                          ].join("\n"),
-                        };
-                      }
-                    }
-                    return {
-                      routeName: "default",
-                      sourceUrl: parsed.toString(),
-                      content: serviceMessage,
-                    };
-                  }
-                  return {
-                    routeName: "default",
-                    sourceUrl: parsed.toString(),
-                    content: defaultResult.stdoutText,
-                  };
-                })();
-            await writeWebFetchCache(cacheKey, fetched);
+        return executeWebFetch(args, {
+          log: buildLogger(client),
+        });
+      },
+    }),
+      [toolIds.websearch]: tool({
+        description: toolDescriptions.websearch,
+      args: {
+        query: tool.schema.string(),
+        category: tool.schema.string().optional(),
+        num_results: tool.schema.number().optional(),
+        offset: tool.schema.number().optional(),
+        recency: tool.schema.number().optional(),
+        domains: tool.schema.array(tool.schema.string()).optional(),
+      },
+      async execute(args, context) {
+        const query = args.query.trim();
+        if (query) {
+          await context.ask({
+            permission: toolIds.websearch,
+            patterns: [query],
+            always: ["*"],
+            metadata: {
+              query,
+            },
+          });
 
-            return formatWebFetchOutput({
-              routeName: fetched.routeName,
-              sourceUrl: fetched.sourceUrl,
-              content: fetched.content,
-            });
-          } catch (error) {
-            const message = error instanceof Error ? error.message : String(error);
-            await client.app.log({
-              body: {
-                service: "web-search-plugin",
-                level: "error",
-                message: "webfetch execution error",
-                extra: {
-                  url: parsed.toString(),
-                  error: message,
-                },
-              },
-            });
-            return [
-              `Tool passphrase: ${PASSPHRASE_WEBFETCH}`,
-              ISSUE_REPORTING_HINT,
-              "Failed to fetch URL.",
-              "If this persists, ask the user to check webfetch/plugin logs and add a report in ISSUES.md.",
-            ].join("\n");
-          }
-        },
-      }),
+          context.metadata({
+            title: `Web search: ${query.slice(0, 72)}`,
+            metadata: {
+              num_results: args.num_results,
+            },
+          });
+        }
 
-      [WEBSEARCH_TOOL_ID]: websearchTool,
+        return executeWebSearch(args, {
+          abort: context.abort,
+          log: buildLogger(client),
+        });
+      },
+    }),
     },
   };
 };
 
-// Backward-compatible export name while transitioning to improved-* naming.
 export const SearxngSearchPlugin = ImprovedWebSearchPlugin;
